% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_when_did_that_happen.R
\name{summarise_dates}
\alias{summarise_dates}
\title{Summarises a vector of dates for each date x identifier}
\usage{
summarise_dates(data, identifier, outcome, dates, FUN = c("min", "max"))
}
\arguments{
\item{data}{(Dataframe) The input data.}

\item{identifier}{(Character) The name of one column in \code{data} that identifies subjects (e.g. patient number or machine serial number).}

\item{outcome}{(Character) The outcome being summarised.}

\item{dates}{(Character) The names of one or more Date/Datetime columns in \code{data} that are related to the same outcome.}

\item{FUN}{(Character) The name of a function to call on \code{dates}. Currently accepts \code{"min"} (to find the earliest date per identifier), and \code{"max"} (the latest date).}
}
\value{
A long-shaped dataframe, with a row for each identifier x date.
}
\description{
Summarises a vector of dates for each date x identifier
}
\examples{
WhenDidThatHappen:::summarise_dates(
  data       = example_events_multirow,
  identifier = "personid",
  outcome    = "Cardiac intervention",
  dates      = c("cied_date", "ablation_date"),
  FUN        = "max"
)

}
