% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_when_did_that_happen.R
\name{when_did_that_happen}
\alias{when_did_that_happen}
\title{Calculate time-to-event and outcomes for simple, composite, and competing events}
\usage{
when_did_that_happen(
  data,
  identifier,
  description,
  start_time,
  event_times,
  early_censors,
  late_censors = NULL,
  time_units = lubridate::days(1),
  blanking = NULL,
  debug = FALSE
)
}
\arguments{
\item{data}{(Dataframe) A dataframe.}

\item{identifier}{(Character) The name of one column in \code{data} that identifies subjects (e.g. patient number or machine serial number).}

\item{description}{(Character) A human-readable name for your analysis, which is used give labels to the output variables. A suffix for the final variable names will be generated from this.}

\item{start_time}{(Character) The name of one Date/Datetime column in \code{data} that provides the start date of observation for each subject.}

\item{event_times}{(List with named Characters) Within this List, the names are human-readable labels for the event, and the contents are the names of Date/Datetime columns in \code{data} that are considered for that event. See 'Examples'.}

\item{early_censors}{(Character) The names of one or more Date/Datetime columns in \code{data} that are used to censor a subject as soon as they occur because no more information can be collected about them, such as death or study withdrawal. At least one censor date must be given in either this argument or in \code{late_censors}. If this argument is not needed, leave it as \code{NULL}.}

\item{late_censors}{(Character) The names of one or more Date/Datetime columns in \code{data} that are used to censor a subject when they stop occurring, such as censoring someone at their last known date of clinical contact. At least one censor date must be given in either this argument or in \code{early_censors}. If this argument is not needed, leave it as \code{NULL}.}

\item{time_units}{(Period) A \code{lubridate} Period object that describes the units of time for each time-to-event. By default, times are reported in elapsed decimal days. You may be interested in \code{\link[lubridate:period]{lubridate::hours()}}, \code{\link[lubridate:period]{lubridate::days()}}, \code{\link[lubridate:period]{lubridate::weeks()}}, or \code{\link[lubridate:period]{lubridate::years()}}. Note that there is no function for months because the number of days in a month varies; use weeks instead.}

\item{blanking}{(Period) \code{NULL} (default), or a \code{lubridate} Period object (like the \code{time_units} argument) that describes how long to wait \emph{after} the \code{start_time} before observing events. See 'Details'.}

\item{debug}{(Logical) If \code{FALSE} (default), returns calculated results. If \code{TRUE}, returns diagnostic results. See 'Value'.}
}
\value{
If \code{debug = FALSE} (default), returns a dataframe of results.

If \code{debug = TRUE}, returns a \code{list} with 2 elements:
\itemize{
\item \verb{$Diagnostic} is a dataframe with all relevant dates in long format,
sorted in their final decided order. Use this to inspect the dates
attributed to each subject and how ties are broken.
\item \verb{$Result} is the normal results dataframe.
}
}
\description{
This function calculates the time that elapsed between a start date and the
earliest outcome date, and what that outcome was. It produces output that is
ready for survival analyses with typical R packages.
}
\section{Details}{
\subsection{Terminology}{
\itemize{
\item \strong{Subject} is the subject of study. This might be unique patients, unique
machines in a factory, etc.
\item \strong{Start date} or \strong{index date} is when observations began for a subject.
Each subject usually has a different start date.
\item \strong{Event date} is when an event of interest happens. For example, the
soonest date when a person gets a heart attack after receiving a pacemaker.
\item \strong{Censor date} is when observations end for a subject. Many events can
censor a subject, including the end of the study, or the last contact we
had with the person, or the date they withdrew consent, or the date when
they died.
\item \strong{Blanking period} is a period after the start date during which we ignore
any events of interest that occur, but still allow for censoring.
}
}

\subsection{Dataframe shapes for input}{

This function requires data in wide format, which means that the date for
each event appears in a separate column. It supports both one-row-per-person
and many-rows-per-person data.
\itemize{
\item In one-row-per-person, it is assumed that the variables contain the earliest
known events (i.e. \code{date_of_earliest_heart_attack}).
\itemize{
\item See the included dataset \code{example_events} for an example of this.
}
\item In many-rows-per-person, the variables can contain the date of every known
event (i.e. \code{date_of_heart_attack}). See
\itemize{
\item See the included dataset \code{example_events_multirow} for an example.
}
}
}

\subsection{Early and late censoring}{

To support multi-row datasets, this function handles censor dates in two
different ways.
\itemize{
\item \strong{Early censors} are events that censor a subject as soon as they happen
because no new information can be collected about them, like when they die.
\itemize{
\item In a multi-row dataset, the minimum of these dates is used for each subject.
}
\item \strong{Late censors} are events that censor a subject when they \strong{stop} happening,
such as when they stop coming to follow-up meetings.
\itemize{
\item In a multi-row dataset, the maximum of these dates is used for each subject.
}
}

In a one-row-per-person dataset, this package assumes that you have
pre-processed the censor columns appropriately. In this case, it doesn't
matter where you put the dates in the \code{early_censors} or \code{late_censors}
arguments.
}

\subsection{Tie-breaking}{

Ties are common if you only have dates of events, but not the time they occur;
a person can have a heart attack, receive surgery for it, and pass away all
in one calendar day.

This function breaks ties by first sorting the dates chronologically \strong{and then}
sorting by the columns you provided in \code{event_times}, \code{early_censors}, and
\code{late_censors}, in the order that you supplied them. This means that if an
event and a censor occur at the same time, then the person will be flagged
with the event. For a call like this, for example:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{when_did_that_happen(
  [...]
  event_times  = list(
    "Cardiac Intervention" = c("ablation_date", "cied_date"),
    "CVD Death"            = c("death_date_cvd")
  ),
  early_censors = c("death_date_noncvd"),
  late_censors  = c("followup_date", "end_of_study")
  [...]
)
}\if{html}{\out{</div>}}

Then the sorting method is:
\enumerate{
\item Sort all dates chronologically, then
\item Sub-sort by events (\code{ablation_date}, then \code{cied_date}, then \code{death_date_cvd}),
\item Then sub-sort by early censors (\code{death_date_noncvd}),
\item Finally, sub-sort by late censors (\code{followup_date} and finally \code{end_of_study}).
}
}

\subsection{Blanking}{

In some analyses, a blanking period can be used to ignore events that occur
too early. A common example in the literature is ignoring early atrial
fibrillation recurrence after surgery, as minor recurrences within 8 weeks or
so are currently believed to be transitory and not clinically significant.
This function implements blanking periods as something that ignores \emph{events},
but does not ignore \emph{censoring}; if your blanking period is 3 months, and a
subject gets an event in 1 month and then dies at 2 months, the event will be
ignored but the person will receive an outcome of Censored and a
time-to-event of 2 months.
}
}

\examples{
# `example_events` is an example dataset included with this package.


# Example of a simple event (Ablated or Censored), reported in units of 1 day
# with no blanking period:

when_did_that_happen(
  data          = example_events,
  identifier    = "personid",
  description   = "Ablation",
  start_time    = "index_date",
  event_times   = list(
    "Ablation"  = c("ablation_date")
  ),
  early_censors = c("death_date"),
  late_censors  = c("followup_date", "end_of_study")
)


# Example of a composite event (Ablated or received CIED implant, versus
# Censored), reported in units of 1 month with a 2 month blanking period:

when_did_that_happen(
  data          = example_events,
  identifier    = "personid",
  description   = "Cardiac Intervention",
  start_time    = "index_date",
  event_times   = list(
    "Cardiac Intervention" = c("ablation_date", "cied_date")
  ),
  early_censors = c("death_date"),
  late_censors  = c("followup_date", "end_of_study"),
  time_units    = lubridate::weeks(4),
  blanking      = lubridate::weeks(8)
)


# Example of competing risks (being ablated before dying). Here I don't have
# an early censor date, so I provide `early_censors = NULL`.

when_did_that_happen(
  data          = example_events,
  identifier    = "personid",
  description   = "Ablation cw Death",
  start_time    = "index_date",
  event_times   = list(
    "Ablation" = c("ablation_date"),
    "Death"    = c("death_date")
  ),
  early_censors = NULL,
  late_censors  = c("followup_date", "end_of_study")
)


# Example of competing risks with a composite outcome (receiving a cardiac
# intervention before dying), with a blanking period of 3 months:

when_did_that_happen(
  data          = example_events,
  identifier    = "personid",
  description   = "Cardiac Intervention cw Death",
  start_time    = "index_date",
  event_times   = list(
    "Cardiac Intervention" = c("ablation_date", "cied_date"),
    "Death"                = c("death_date")
  ),
  early_censors = NULL,
  late_censors  = c("followup_date", "end_of_study"),
  time_units    = lubridate::days(1),
  blanking      = lubridate::weeks(12)
)

}
