% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WhenDidThatHappen-package.R
\docType{package}
\name{WhenDidThatHappen-package}
\alias{WhenDidThatHappen}
\alias{WhenDidThatHappen-package}
\title{WhenDidThatHappen: Calculate Outcome and Time-To-Event Inputs For Survival Analyses}
\description{
The goal of 'WhenDidThatHappen' is to take raw Date or Datetime data (either as one-row-per-subject or many-rows-per-subject) and produce outcome and time-to-event outputs that are ready for survival analysis using common packages such as 'survival', 'coxme', and 'cmprsk'. 'WhenDidThatHappen' supports simple and composite outcomes, with optional blanking periods and optional minimum observation periods.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DesiQuintans/WhenDidThatHappen}
  \item Report bugs at \url{https://github.com/DesiQuintans/WhenDidThatHappen/issues}
}

}
\author{
\strong{Maintainer}: Desi Quintans \email{science@desiquintans.com}

}
\keyword{internal}
