% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_data.R
\docType{data}
\name{example_events}
\alias{example_events}
\title{Fictional time-to-event data (one row per person)}
\format{
\subsection{\code{example_events}}{

A data frame with 40 rows and 8 columns:
\describe{
\item{personid}{Person identifier.}
\item{studyarm}{Study arm.}
\item{index_date}{Index date for each person.}
\item{heartsurgery_date}{Date of earliest heart surgery for each person. Missing if it did not happen.}
\item{lungsurgery_date}{Date of earliest lung surgery for each person. Missing if it did not happen.}
\item{death_date}{Date of death for each person. Missing if they did not die.}
\item{followup_date}{Last contact date for each person. Missing if they were only contacted at the beginning of the study.}
\item{readmit_date}{Readmission date for each person. Missing if they were not readmitted.}
\item{end_of_study}{Date of the end of the study.}
...
}
}
}
\usage{
example_events
}
\description{
An invented dataset for demonstrating and testing time-to-event calculations.
}
\keyword{datasets}
