% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_data.R
\docType{data}
\name{example_events_multirow}
\alias{example_events_multirow}
\title{Fictional time-to-event data (many rows per person)}
\format{
\subsection{\code{example_events_multirow}}{

A data frame with 90 rows and 8 columns:
\describe{
\item{personid}{Person identifier.}
\item{index_date}{Index date for each person.}
\item{ablation_date}{Dates of ablation for each person. All missing if it did not happen.}
\item{cied_date}{Dates of cardiac implanted electronic device implantation/revision for each person. All missing if it did not happen.}
\item{death_date}{Date of death for each person. All missing if they did not die.}
\item{followup_date}{All dates of clinical contact for each person. All missing if they were only contacted at the beginning of the study.}
\item{readmit_date}{Readmission dates for each person. All missing if they were never readmitted.}
\item{end_of_study}{Date of the end of the study.}
...
}
}
}
\usage{
example_events_multirow
}
\description{
An invented dataset for demonstrating and testing time-to-event calculations.
Your dataset will probably be cleaner!
}
\keyword{datasets}
