% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_when_did_that_happen.R
\name{when_did_that_happen}
\alias{when_did_that_happen}
\title{Calculate time-to-event and outcomes for simple, composite, and competing events}
\usage{
when_did_that_happen(
  data,
  analysis,
  identifier,
  start_time,
  event_times,
  early_censors = NULL,
  late_censors = NULL,
  time_units = lubridate::days(1),
  blanking = lubridate::days(0),
  minimum_time = lubridate::days(0),
  debug = FALSE
)
}
\arguments{
\item{data}{(Dataframe) The input data.}

\item{analysis}{(Character) A human-readable name for your analysis, which is used give labels to the output variables. A suffix for the final variable names will be generated from this.}

\item{identifier}{(Character) The name of one column in \code{data} that identifies subjects (e.g. patient number or machine serial number).}

\item{start_time}{(Character) The name of one Date/Datetime column in \code{data} that provides the start date of observation for each subject.}

\item{event_times}{(List with named Characters) Within this List, the names are human-readable labels for the event, and the contents are the names of Date/Datetime columns in \code{data} that are considered for that event. See 'Examples'.}

\item{early_censors}{(Character) The names of one or more Date/Datetime columns in \code{data} that are used to censor a subject as soon as they occur because no more information can be collected about them, such as death or study withdrawal. At least one censor date must be given in either this argument or in \code{late_censors}. If this argument is not needed, leave it as \code{NULL}.}

\item{late_censors}{(Character) The names of one or more Date/Datetime columns in \code{data} that are used to censor a subject when they stop occurring, such as censoring someone at their last known date of clinical contact. At least one censor date must be given in either this argument or in \code{early_censors}. If this argument is not needed, leave it as \code{NULL}.}

\item{time_units}{(Period) A \code{lubridate} Period object that describes the units of time for each time-to-event. By default, times are reported in elapsed decimal days. You may be interested in \code{\link[lubridate:period]{lubridate::hours()}}, \code{\link[lubridate:period]{lubridate::days()}}, \code{\link[lubridate:period]{lubridate::weeks()}}, or \code{\link[lubridate:period]{lubridate::years()}}. Note that there is no function for months because the number of days in a month varies; use weeks instead.}

\item{blanking}{(Period) A \code{lubridate} Period object (like the \code{time_units} argument) that describes how long to wait \emph{after} the \code{start_time} before observing events.}

\item{minimum_time}{(Period) A \code{lubridate} Period object that describes the minimum follow-up time required for a subject to be eligible for this analysis (they will be \code{NA} if ineligible).}

\item{debug}{(Logical) If \code{FALSE} (default), returns calculated results. If \code{TRUE}, returns diagnostic results. See 'Value'.}
}
\value{
If \code{debug = FALSE} (default), returns a dataframe of results.

If \code{debug = TRUE}, returns a \code{list} with 2 elements:
\itemize{
\item \verb{$Diagnostic} is a dataframe with all relevant dates in long format,
sorted in their final decided order. Use this to inspect the dates
attributed to each subject and how ties are broken.
\item \verb{$Result} is the normal results dataframe.
}
}
\description{
This function calculates the time that elapsed between a start date and the
earliest outcome date, and what that outcome was. It produces output that is
ready for survival analyses with typical R packages.
}
\details{
\subsection{Dataframe shapes for input}{

This function requires data in wide format, which means that the date for
each event appears in a separate column. It supports both one-row-per-person
and many-rows-per-person data.
\itemize{
\item In one-row-per-person, it is assumed that the variables contain the earliest
known events (i.e. \code{date_of_earliest_heart_attack}).
\itemize{
\item See the included dataset \code{example_events} for an example of this.
}
\item In many-rows-per-person, the variables can contain the date of every known
event (i.e. \code{date_of_heart_attack}). See
\itemize{
\item See the included dataset \code{example_events_multirow} for an example.
}
}
}
}
\examples{
# `example_events` is an example dataset included with this package.


# Example of a simple event (Ablated or Censored), reported in units of 1 day
# with no blanking period or minimum follow-up requirement:

when_did_that_happen(
  data          = example_events,
  analysis      = "Ablation",

  identifier    = "personid",
  start_time    = "index_date",
  event_times   = list(
    "Ablation"  = c("ablation_date")
  ),
  early_censors = c("death_date", "end_of_study"),
  late_censors  = c("followup_date"),

  time_units    = lubridate::days(1),
  blanking      = lubridate::days(0),
  minimum_time  = lubridate::days(0),

  debug = FALSE
)


# Example of a composite event (Ablated or received CIED implant, versus
# Censored), reported in units of 1 month with a 2 month blanking period:

when_did_that_happen(
  data          = example_events,
  analysis      = "Cardiac Intervention",

  identifier    = "personid",
  start_time    = "index_date",
  event_times   = list(
    "Cardiac Intervention" = c("ablation_date", "cied_date")
  ),
  early_censors = c("death_date", "end_of_study"),
  late_censors  = c("followup_date"),

  time_units    = lubridate::weeks(4),
  blanking      = lubridate::weeks(8),
  minimum_time  = lubridate::days(0),

  debug = FALSE
)


# Example of competing risks with a composite outcome (receiving a cardiac
# intervention) versus a simple outcome (dying), with a blanking period
# of 1 month and a requirement that people must have at least 6 months of
# observation time in the study.

when_did_that_happen(
  data          = example_events,
  analysis      = "Cardiac Intervention cw Death",

  identifier    = "personid",
  start_time    = "index_date",
  event_times   = list(
    "Cardiac Intervention" = c("ablation_date", "cied_date"),
    "Death"                = c("death_date")
  ),
  early_censors = c("end_of_study"),
  late_censors  = c("followup_date"),

  time_units    = lubridate::days(1),
  blanking      = lubridate::weeks(4),
  minimum_time  = lubridate::weeks(24),

  debug = FALSE
)

}
