% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WhenDidThatHappen-package.R
\docType{package}
\name{WhenDidThatHappen-package}
\alias{WhenDidThatHappen}
\alias{WhenDidThatHappen-package}
\title{WhenDidThatHappen: Calculate Outcome and Time-To-Event Inputs For Survival Analyses}
\description{
'WhenDidThatHappen' prepares survival analyses. It takes your raw Datetimes (either as one- or many-rows-per-subject) and derives time-to-event variables for use with common packages such as 'survival', 'coxme', and 'cmprsk'. It supports right-censored simple and composite outcomes, with optional blanking periods and minimum observation periods.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DesiQuintans/WhenDidThatHappen}
  \item Report bugs at \url{https://github.com/DesiQuintans/WhenDidThatHappen/issues}
}

}
\author{
\strong{Maintainer}: Desi Quintans \email{science@desiquintans.com}

}
\keyword{internal}
